# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TheDotFactoryChanges - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://localhost:{port}/api/{basePath}
    variables:
      port:
        enum:
          - '59572'
        default: '59572'
      basePath:
        enum:
          - 'v1'
        default: v1
paths:
  /users:
    get:
      tags:
        - user
      summary: Get users information
      description: Get users information
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Not allowed
        '500':
          description: Server error
  /users/{userId}:
    get:
      tags:
        - user
      summary: Get user information
      description: Get user information
      parameters:
        - name: userId
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '403':
          description: Not allowed
        '404':
          description: User not found
        '500':
          description: Server error
    patch:
      tags:
        - user
      summary: Get user information
      description: Get user information
      parameters:
        - name: userId
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User registration
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '403':
          description: Not allowed
        '404':
          description: User not found
        '500':
          description: Server error
  /users/login:
    get:
      tags:
        - user
      summary: User login
      description: User login
      parameters:
        - name: login
          in: query
          description: Login name
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '500':
          description: Server error
  /users/logout:
    get:
      tags:
        - user
      summary: User logout
      description: User logout
      parameters:
        - name: id
          in: query
          description: The user id for logout
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '403':
          description: Not allowed
        '500':
          description: Server error
  /users/registration:
    post:
      tags:
        - user
      summary: User registration
      description: User registration
      requestBody:
        description: User registration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '400':
          description: User is already registered
        '500':
          description: Server error
...
components:
  schemas:
    User:
      required:
        - name
        - login
        - password
      type: object
      properties:
        name:
          type: string
          example: string
        login:
          type: string
          example: string
        password:
          type: string
          example: string
    UserPatchRequest:
      required:
        - op
        - value
        - path
      type: object
      properties:
        op:
          description: Operation
          type: string
          enum:
            - ChangePassword
            - ChangeName
        path:
          description: A JSON Pointer path.
          type: string
        value:
          description: The value to add, replace or test.
          type: string
    NewUser:
      required:
        - login
        - password
        - repPassword
      type: object
      properties:
        login:
          type: string
          example: string
        password:
          type: string
          example: string
        repPassword:
          type: string
          example: string
...